requirements.txt
flask>=3.1.3
psycopg2-binary>=2.9.11
python-dotenv>=1.0.0
